Move logging utilities to 'logging' package
